requirements.txt
PyQt6
mysql-connector
mysql-connector-python
python-dotenv
pymysql